reuse query parameter parsing refactors model into separate files too
