+++ components/fms-server/src/main.rs
@@ -27,15 +27,19 @@ use log::{error, info};
 use serde_json::json;
 use std::collections::HashMap;
 use std::process;
-use std::str::FromStr;
 use std::sync::Arc;
 
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 
 use influx_reader::InfluxReader;
 
 mod influx_reader;
 mod models;
+mod query_parser;
+
+use query_parser::parse_query_parameters;
+use models::position::{VehiclePositionResponseObject, VehiclePositionResponseObjectVehiclePositionResponse};
+use models::status::{VehicleStatusResponseObjectVehicleStatusResponse, VehicleStatusResponseObject};
 
 #[tokio::main]
 async fn main() {
@@ -71,29 +75,15 @@ async fn get_vehicleposition(
     State(influx_server): State<Arc<InfluxReader>>,
     Query(params): Query<HashMap<String, String>>,
 ) -> Result<Json<serde_json::Value>, StatusCode> {
-    let start_time = params.get("starttime").map_or(0, |text| {
-        DateTime::<Utc>::from_str(text).map_or(0, |time| time.timestamp())
-    });
-
-    let stop_time = params
-        .get("stoptime")
-        .map_or_else(|| Utc::now().timestamp(), |text| {
-            DateTime::<Utc>::from_str(text).map_or_else(|_| Utc::now().timestamp(), |time| time.timestamp())
-        });
-
-    let vin = params.get("vin");
-    let trigger_filter = params.get("triggerFilter");
-    let latest_only = params
-        .get("latestOnly")
-        .map_or(false, |text| text.parse().unwrap_or(false));
+    let query_parameters = parse_query_parameters(&params)?;
 
     influx_server
-        .get_vehicleposition(start_time, stop_time, vin, trigger_filter, latest_only)
+        .get_vehicleposition(&query_parameters)
         .await
         .map(|positions| {
-            let result = json!(models::VehiclePositionResponseObject {
+            let result = json!(VehiclePositionResponseObject {
                 vehicle_position_response:
-                    models::VehiclePositionResponseObjectVehiclePositionResponse {
+                    VehiclePositionResponseObjectVehiclePositionResponse {
                         vehicle_positions: Some(positions)
                     },
                 more_data_available: false,
@@ -117,11 +107,11 @@ async fn get_vehicles(
         .get_vehicles()
         .await
         .map(|vehicles| {
-            let response = models::VehicleResponseObjectVehicleResponse {
+            let response = models::vehicle::VehicleResponseObjectVehicleResponse {
                 vehicles: Some(vehicles),
             };
 
-            let result_object = json!(models::VehicleResponseObject {
+            let result_object = json!(models::vehicle::VehicleResponseObject {
                 vehicle_response: response,
                 more_data_available: false,
                 more_data_available_link: None,
@@ -134,67 +124,23 @@ async fn get_vehicles(
         })
 }
 
-fn parse_latest_only(params: &HashMap<String, String>) -> Result<Option<bool>, StatusCode> {
-    let latest_parameter = params.get("latestOnly");
-    if let Some(latest_string) = latest_parameter {
-        let latest_result = latest_string.parse();
-        if latest_result.is_err() {
-            return Err(StatusCode::BAD_REQUEST);
-        }
-        return Ok(latest_result.ok());
-    }
-    Ok(None)
-}
-
 async fn get_vehiclesstatuses(
     State(influx_server): State<Arc<InfluxReader>>,
     Query(params): Query<HashMap<String, String>>,
 ) -> Result<Json<serde_json::Value>, StatusCode> {
     
-    
-    let start_parameter = params.get("starttime");
-    let stop_parameter = params.get("stoptime");
-    
-    
-    let latest_only = match parse_latest_only(&params) {
-        Ok(value) => value,
-        Err(status) => return Err(status), 
-    };
-    
-    if start_parameter.is_none() && latest_only.is_none() {
-        // rFMS makes it mandatory to either supply the starttime or latestOnly
-        return Err(StatusCode::BAD_REQUEST);
-    }
-
-    if latest_only.is_some() && (start_parameter.is_some() || stop_parameter.is_some()) {
-        // rFMS does not allow to set latestOnly and and time at the same time
-        return  Err(StatusCode::BAD_REQUEST);
-    }
-
-    let start_time = start_parameter.map_or(0, |text| {
-        DateTime::<Utc>::from_str(text).map_or(0, |time| time.timestamp())
-    });
-
-    let stop_time = stop_parameter
-        .map_or_else(|| Utc::now().timestamp(), |text| {
-            DateTime::<Utc>::from_str(text).map_or_else(|_| Utc::now().timestamp(), |time| time.timestamp())
-        });
-
-    let vin = params.get("vin");
-    let trigger_filter = params.get("triggerFilter");
-
-
+    let query_parameters = parse_query_parameters(&params)?;
     influx_server
-        .get_vehiclesstatuses(start_time, stop_time, vin, trigger_filter, latest_only)
+        .get_vehiclesstatuses(&query_parameters)
         .await
         .map(|vehicles_statuses| {
-            let response = models::VehicleStatusResponseObjectVehicleStatusResponse {
+            let response = VehicleStatusResponseObjectVehicleStatusResponse {
                 vehicle_statuses: Some(vehicles_statuses),
             };
 
             //TODO for request_server_date_time
             // put in start time used in influx query instead of now
-            let result_object = json!(models::VehicleStatusResponseObject {
+            let result_object = json!(VehicleStatusResponseObject {
                 vehicle_status_response: response,
                 more_data_available: false,
                 more_data_available_link: None,
